wip: fixed unit test imports following type move
