+++ src/app/forms/components/text-input/text-input.component.spec.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
-import { FormControl, FormGroup } from '@angular/forms';
-import { CustomFormControl, FormNodeTypes } from '../../services/dynamic-form.service';
+import { FormGroup } from '@angular/forms';
+import { CustomFormControl, FormNodeTypes } from '../../services/dynamic-form.types';
 import { TextInputComponent } from './text-input.component';
 
 @Component({
