+++ src/app/forms/components/view-list-item/view-list-item.component.spec.ts
@@ -1,7 +1,7 @@
 import { Component } from '@angular/core';
 import { ComponentFixture, TestBed } from '@angular/core/testing';
 import { FormGroup } from '@angular/forms';
-import { CustomFormControl, FormNodeTypes } from '../../services/dynamic-form.service';
+import { CustomFormControl, FormNodeTypes } from '../../services/dynamic-form.types';
 import { ViewListItemComponent } from './view-list-item.component';
 
 @Component({
